[FIX] edrv-8139: Fix compiler warning
Fixes:
edrv-8111.c:1765:5: warning: format '%lu' expects argument of type 'long unsigned int', but argument 2 has type 'unsigned int' [-Wformat=]
     printk("enable DMA- size: %lu\n", sizeof(dma_addr_t));
     ^

+++ stack/src/kernel/edrv/edrv-8111.c
@@ -1762,7 +1762,7 @@ static INT initOnePciDev(struct pci_dev* pPciDev_p,
     }
 
     // Set Dma mask size to 32 bits
-    printk("enable DMA- size: %lu\n", sizeof(dma_addr_t));
+    printk("enable DMA- size: %lu\n", (ULONG)sizeof(dma_addr_t));
     result = pci_set_dma_mask(pPciDev_p, DMA_BIT_MASK(32));
     if (result != 0)
     {
